+++ services-js/registry-certs/client/marriageintention/forms/partnerFormB.tsx
@@ -216,6 +216,8 @@ export default class PartnerForm extends Component<Props> {
                 val: e.target.value,
                 partnerFlag: 'B',
                 certObj: this.props.marriageIntentionCertificateRequest,
+                requestInformation:
+                  marriageIntentionCertificateRequest.requestInformation,
               }),
             formErrors: this.props.formErrors,
             errorElemSrc: this.props.errorElemSrc,
